requirements.txt
ccxt==4.4.26
python-dotenv==1.0.1
streamlit==1.41.1
plotly==5.24.1
pandas==2.2.3
